@@ -15,14 +15,14 @@
         }
     },
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=5.4.0"
     },
     "require-dev": {
+        "phpunit/phpunit": "4.8.*",
         "symfony/stopwatch": "~2.3|~3.0",
         "myclabs/php-enum": "~1.5",
         "marc-mabe/php-enum": "~2.3",
         "happy-types/enumerable-type": "~1.0",
-        "phpunit/phpunit": "4.8.*",
         "scrutinizer/ocular": "~1.3",
         "satooshi/php-coveralls": "^1.0"
     }
